Added base playbook
Installs Jetty and Shibboleth IdP.

Broken into two tasks:
Jetty:
- Downloads a Jetty 9.2 distribution (officially supported by Shibboleth)
- Creates jetty user
- Sets basic Jetty configuration (jetty.xml / modules etc)
- Register service w/ runlevels
IdP:
- Downloads Shibboleth IdP 3.1.1
- Create a new JETTY_BASE to run Shibboleth IdP. See
  http://www.eclipse.org/jetty/documentation/current/startup-base-and-home.html
- Runs Shibboleth IdP installer with pexpect
- Downloads placeholder keystore (Jetty fails if this not present when using the
  ssl module (required by Shibboleth IdP)

At this point the Shibboleth IdP is NOT configured for the federation. However
Jetty should start without failure (after running `service jetty start`). To
verify you should get a valid response when curling
http://localhost:8080/idp/status.

This playbook was designed with idempotency in mind and was tested with a fresh
Centos 7 installation.