Get the Metadata certificate verified and not verified
As we can not inject the sha256sum value for the certificate
into an existing IdP config we need to allow existing IdPs
to get the metadata certificate without any sha256sum verfication.

New IdP will have the sha256sum and can thus verify the download
of the signing certificate.