Removed the blank metadata file.
IdP v3.3.0 does not just ignore this file, it attempts to process
it and fails. It does not then attempt to load a new version from
its http source.

As the IdP can now write to this directory due to an earlier
resovled issue the the can just load the metadata file itself
without the need for a place marker file.